requirements.txt
beautifulsoup4~=4.13.4
requests~=2.32.3
flask~=3.1.0
functions-framework
google-cloud-storage
scikit-learn
pandas~=2.2.3
matplotlib~=3.10.3
yfinance~=0.2.61
streamlit
pytest~=8.4.2
